refactor: deduplicate sample HTML, fix defects, rename OnlyAtServer projects
- Extracted the ~90-line HTML fixture duplicated identically across
  ClientSendHttp, ClientSendTcp, OnlyAtServer v1/v2 into a single
  trimmed version: fixed the malformed <a href="file:./myfile.txt</a>
  anchor, dropped the via.placeholder.com network dependency, and
  removed the ANSI-color <pre> block that added noise with no
  didactic value.
- ClientSendTcp: removed a dead AddHttpClient("HtmlPdfServer", ...) +
  IHttpClientFactory registration left over from copy-pasting the HTTP
  sample (never used - the sample sends via SendToTcpServer over TCP),
  fixed a console message that claimed "via http post" on a TCP-only
  sample, and lowered Timeout(500000) to a sane 30s matching the
  sample's own TCP wait timeout.
- Renamed OnlyAtServer v1/v2 to CustomHooks/QuickStart so the project
  name says what it demonstrates (BeforePDF/AfterPDF hooks + typed I/O
  vs. minimal default byte[] usage), updated the .sln and README links
  to match.
- Fixed the stale src/README.txt server-side snippet that hand-rolled
  [FromBody] byte[] - now incompatible with the client after the wire
  contract fix - to use MapHtmlPdfEndpoints() instead.
