+++ samples/ConsoleHtmlToPdfPlus.ClientSendTcp/Program.cs
@@ -46,11 +46,8 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
             //token to gracefull shutdown
             var applifetime = HostApp.Services.GetService<IHostApplicationLifetime>()!;
 
-            //client http to endpoint    
-            var clienthttp = HostApp!.Services.GetRequiredService<IHttpClientFactory>().CreateClient("HtmlPdfServer");
-
-            //create client instance and to HtmlPdfPlus server endpoint
-            Console.WriteLine($"HtmlPdfClient send Html to PDF Server via http post");
+            //create client instance and send to HtmlPdfPlus server via the custom TCP transport below
+            Console.WriteLine($"HtmlPdfClient send Html to PDF Server via TCP");
 
             var pdfresult = await HtmlPdfClient.Create("HtmlPdfPlusClient")
                              .PageConfig((cfg) =>
@@ -63,7 +60,7 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                              })
                              .Logger(HostApp.Services.GetService<ILogger<Program>>())
                              .FromHtml(HtmlSample())
-                             .Timeout(500000)
+                             .Timeout(30000)
                              .Run(SendToTcpServer, applifetime.ApplicationStopping);
 
             Console.WriteLine($"HtmlPdfClient IsSuccess {pdfresult.IsSuccess} after {pdfresult.ElapsedTime}");
@@ -152,13 +149,6 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                         .AddFilter("Microsoft", LogLevel.Warning)
                         .AddFilter("System", LogLevel.Warning)
                         .AddConsole();
-                })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    services.AddHttpClient("HtmlPdfServer", httpClient =>
-                    {
-                        httpClient.BaseAddress = new Uri("https://localhost:7212/GeneratePdf");
-                    });
                 });
 
         private static string HtmlSample()
@@ -254,7 +244,7 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                     <nav>
                         <a href="https://www.microsoft.com">Microsoft</a>
                         <a href="https://www.google.com">Google</a>
-                        <a href="file:./myfile.txt</a>
+                        <a href="https://github.com/FRACerqueira/HtmlPdfPlus">HtmlPdfPlus</a>
                     </nav>
                     <section>
                         <article>
@@ -272,21 +262,6 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                             <p>Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
                             <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/w8AAn8C4iO5fwAAAABJRU5ErkJggg==" alt="Black dot" class="base64-image">
                         </article>
-                        <article>
-                        <h2>Article 4</h2>
-                            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque facilisis.</p>
-                            <img src="https://via.placeholder.com/300" alt="Placeholder Image" class="base64-image">
-                        </article>
-                            <article>
-                                <h2>Article 5</h2>
-                                <p>Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.</p>
-                                <img src="https://via.placeholder.com/150" alt="Placeholder Image" class="base64-image">
-                            </article>
-                            <article>
-                                <h2>Article 6</h2>
-                                <p>Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.</p>
-                                <img src="https://via.placeholder.com/100" alt="Placeholder Image" class="base64-image">
-                            </article>
                     </section>
                     <section>
                         <h2>Colors</h2>
@@ -301,13 +276,13 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                         <form action="#" method="post">
                             <label for="name">Name:</label>
                             <input type="text" id="name" name="name" placeholder="Enter your name">
-    
+
                             <label for="email">Email:</label>
                             <input type="text" id="email" name="email" placeholder="Enter your email">
-    
+
                             <label for="message">Message:</label>
                             <textarea id="message" name="message" rows="5" placeholder="Your message..."></textarea>
-    
+
                             <label for="rating">Rating:</label>
                             <select id="rating" name="rating">
                                 <option value="excellent">Excellent</option>
@@ -315,34 +290,10 @@ namespace ConsoleHtmlToPdfPlus.ClientSendTcp
                                 <option value="average">Average</option>
                                 <option value="poor">Poor</option>
                             </select>
-    
+
                             <button type="submit">Submit</button>
                         </form>
                     </section>
-        
-                <br />
-    
-                    <pre>
-                                         40m     41m     42m     43m     44m     45m     46m     47m
-                                <span style="">  gYw  </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; background-color: white; ">  gYw  </span>
-                        <span style="">    1m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; ">  gYw   </span><span style="font-weight: bold; "></span><span style="font-weight: bold; background-color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; "></span><span style="font-weight: bold; background-color: white; ">  gYw  </span>
-                        <span style="">   30m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; ">  gYw   </span><span style="color: dimgray; "></span><span style="background-color: dimgray; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: red; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: lime; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: yellow; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: #3333FF; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: fuchsia; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: aqua; color: dimgray; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="background-color: white; color: dimgray; ">  gYw  </span>
-                        <span style=""> 1;30m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; ">  gYw   </span><span style="font-weight: bold; color: dimgray; "></span><span style="font-weight: bold; background-color: dimgray; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: red; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: lime; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: yellow; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: #3333FF; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: fuchsia; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: aqua; color: dimgray; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: dimgray; "></span><span style="font-weight: bold; background-color: white; color: dimgray; ">  gYw  </span>
-                        <span style="">   31m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; ">  gYw   </span><span style="color: red; "></span><span style="background-color: dimgray; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: red; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: lime; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: yellow; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: #3333FF; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: fuchsia; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: aqua; color: red; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="background-color: white; color: red; ">  gYw  </span>
-                        <span style=""> 1;31m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; ">  gYw   </span><span style="font-weight: bold; color: red; "></span><span style="font-weight: bold; background-color: dimgray; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: red; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: lime; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: yellow; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: #3333FF; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: fuchsia; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: aqua; color: red; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: red; "></span><span style="font-weight: bold; background-color: white; color: red; ">  gYw  </span>
-                        <span style="">   32m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; ">  gYw   </span><span style="color: lime; "></span><span style="background-color: dimgray; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: red; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: lime; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: yellow; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: #3333FF; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: fuchsia; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: aqua; color: lime; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="background-color: white; color: lime; ">  gYw  </span>
-                        <span style=""> 1;32m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; ">  gYw   </span><span style="font-weight: bold; color: lime; "></span><span style="font-weight: bold; background-color: dimgray; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: red; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: lime; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: yellow; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: #3333FF; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: fuchsia; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: aqua; color: lime; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: lime; "></span><span style="font-weight: bold; background-color: white; color: lime; ">  gYw  </span>
-                        <span style="">   33m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; ">  gYw   </span><span style="color: yellow; "></span><span style="background-color: dimgray; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: red; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: lime; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: yellow; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: #3333FF; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: fuchsia; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: aqua; color: yellow; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="background-color: white; color: yellow; ">  gYw  </span>
-                        <span style=""> 1;33m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; ">  gYw   </span><span style="font-weight: bold; color: yellow; "></span><span style="font-weight: bold; background-color: dimgray; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: red; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: lime; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: yellow; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: #3333FF; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: fuchsia; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: aqua; color: yellow; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: yellow; "></span><span style="font-weight: bold; background-color: white; color: yellow; ">  gYw  </span>
-                        <span style="">   34m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; ">  gYw   </span><span style="color: #3333FF; "></span><span style="background-color: dimgray; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: red; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: lime; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: yellow; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: #3333FF; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: fuchsia; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: aqua; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="background-color: white; color: #3333FF; ">  gYw  </span>
-                        <span style=""> 1;34m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; ">  gYw   </span><span style="font-weight: bold; color: #3333FF; "></span><span style="font-weight: bold; background-color: dimgray; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: red; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: lime; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: yellow; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: #3333FF; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: fuchsia; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: aqua; color: #3333FF; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: #3333FF; "></span><span style="font-weight: bold; background-color: white; color: #3333FF; ">  gYw  </span>
-                        <span style="">   35m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; ">  gYw   </span><span style="color: fuchsia; "></span><span style="background-color: dimgray; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: red; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: lime; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: yellow; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: #3333FF; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: fuchsia; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: aqua; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="background-color: white; color: fuchsia; ">  gYw  </span>
-                        <span style=""> 1;35m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; ">  gYw   </span><span style="font-weight: bold; color: fuchsia; "></span><span style="font-weight: bold; background-color: dimgray; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: red; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: lime; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: yellow; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: #3333FF; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: fuchsia; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: aqua; color: fuchsia; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: fuchsia; "></span><span style="font-weight: bold; background-color: white; color: fuchsia; ">  gYw  </span>
-                        <span style="">   36m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; ">  gYw   </span><span style="color: aqua; "></span><span style="background-color: dimgray; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: red; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: lime; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: yellow; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: #3333FF; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: fuchsia; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: aqua; color: aqua; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="background-color: white; color: aqua; ">  gYw  </span>
-                        <span style=""> 1;36m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; ">  gYw   </span><span style="font-weight: bold; color: aqua; "></span><span style="font-weight: bold; background-color: dimgray; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: red; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: lime; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: yellow; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: #3333FF; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: fuchsia; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: aqua; color: aqua; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: aqua; "></span><span style="font-weight: bold; background-color: white; color: aqua; ">  gYw  </span>
-                        <span style="">   37m </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; ">  gYw   </span><span style="color: white; "></span><span style="background-color: dimgray; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: red; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: lime; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: yellow; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: #3333FF; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: fuchsia; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: aqua; color: white; ">  gYw  </span><span style=""> </span><span style="background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="background-color: white; color: white; ">  gYw  </span>
-                        <span style=""> 1;37m </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; ">  gYw   </span><span style="font-weight: bold; color: white; "></span><span style="font-weight: bold; background-color: dimgray; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: red; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: lime; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: yellow; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: #3333FF; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: fuchsia; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: aqua; color: white; ">  gYw  </span><span style=""> </span><span style="font-weight: bold; background-color: initial; color: initial; font-weight: normal; opacity: 1.0; font-style: normal; text-decoration: none; display: inline; color: white; "></span><span style="font-weight: bold; background-color: white; color: white; ">  gYw  </span>
-                    </pre>
         
                     <footer>
                         <p>&copy; 2024 HTML to PDF Test Page</p>
