+++ samples/ConsoleHtmlToPdfPlus.OnlyAtServer/CustomHooks/ConsoleHtmlToPdfPlus.OnlyAtServerCustomHooks.csproj
