+++ samples/ConsoleHtmlToPdfPlus.OnlyAtServer/CustomHooks/Program.cs
@@ -9,133 +9,64 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace ConsoleHtmlToPdfPlus.ClientSendHttp
+namespace ConsoleHtmlToPdfPlus.OnlyAtServerCustomHooks
 {
     public class Program
     {
         private static readonly string PathToSamples = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
-        private static IHost? HostApp = null;
+
         public static async Task Main(string[] args)
         {
-            Console.WriteLine("Example of HTML to PDF Plus console using only Client with all settings sent via http to server PDF");
-            Console.WriteLine("===================================================================================================");
-            Console.WriteLine(""); 
-            Console.WriteLine("Start the WebHtmlToPdf.GenericServer Server project first. When ready, press any key to continue");
-            Console.WriteLine("");
-            Console.ReadKey();
+            Console.WriteLine("Example: server-only conversion with BeforePDF/AfterPDF hooks (token substitution, custom file output) and DisableCompress for same-process performance");
+            Console.WriteLine("=====================================================================================================================================================");
 
-            HostApp = CreateHostBuilder(args).Build();
+            var HostApp = CreateHostBuilder(args).Build();
 
-             //token to gracefull shutdown
+            Console.WriteLine("Warmup HtmlPdfServerPlus with buffer");
+
+            //Warmup HtmlPdfServerPlus on startup for better performance from the first request
+            var WarmupTS = HostApp.WarmupHtmlPdfService<string, string>();
+            Console.WriteLine($"HtmlPdfServerPlus ready after {WarmupTS}");
+
+            //token to gracefull shutdown
             var applifetime = HostApp.Services.GetService<IHostApplicationLifetime>()!;
 
-            //client http to endpoint    
-            var clienthttp = HostApp!.Services.GetRequiredService<IHttpClientFactory>().CreateClient("HtmlPdfServer");
+            //instance of Html to Pdf Engine
+            var PDFserver = HostApp!.Services.GetHtmlPdfService<string, string>();
 
-            //create client instance and to HtmlPdfPlus server endpoint
-            Console.WriteLine($"HtmlPdfClient send Html to PDF Server via http post");
+            //Performs conversion and custom operations on the server
+            var pdfresult = await PDFserver
+                .ScopeData(Path.Combine(PathToSamples, "html2pdfHtml.pdf"))
+                .FromHtml(HtmlSample(), 5000)
+                .BeforePDF((html, _, _) =>
+                {
+                    //performs replacement token substitution in the HTML source before performing the conversion
+                    var aux = html.Replace("[{MyTokenTemplate}]", "HTML to PDF Test");
+                    return Task.FromResult(aux);
+                })
+                .AfterPDF(async (pdfbyte, filepath, token) =>
+                {
+                    //performs writing to file after performing conversion
+                    await File.WriteAllBytesAsync(filepath!, pdfbyte!, token);
+                    Console.WriteLine($"File PDF generate at {filepath}");
+                    return filepath!;
+                })
+                .Run(applifetime.ApplicationStopping);
 
-            var pdfresult = await HtmlPdfClient.Create("HtmlPdfPlusClient")
-                             .PageConfig((cfg) =>
-                             {
-                                 cfg.Margins(10)
-                                   .Footer("'<span style=\"text-align: center;width: 100%;font-size: 10px\"> <span class=\"pageNumber\"></span> of <span class=\"totalPages\"></span></span>")
-                                   .Header("'<span style=\"text-align: center;width: 100%;font-size: 10px\" class=\"title\"></span>")
-                                   .Orientation(PageOrientation.Landscape)
-                                   .DisplayHeaderFooter(true);
-                             })
-                             .Logger(HostApp.Services.GetService<ILogger<Program>>())
-                             .FromHtml(HtmlSample())
-                             .Timeout(5000)
-                             .Run(clienthttp, applifetime.ApplicationStopping);
+            Console.WriteLine($"HtmlPdfServer IsSuccess {pdfresult.IsSuccess} after {pdfresult.ElapsedTime}");
 
-            Console.WriteLine($"HtmlPdfClient IsSuccess {pdfresult.IsSuccess} after {pdfresult.ElapsedTime}");
-
-            //performs writing to file after performing conversion
             if (pdfresult.IsSuccess)
             {
-                var fullpath = Path.Combine(PathToSamples, "html2pdfHtml.pdf");
-                await File.WriteAllBytesAsync(fullpath, pdfresult.OutputData!);
-                Console.WriteLine($"File PDF generate at {fullpath}");
+                Console.WriteLine($"File PDF generate at {pdfresult.OutputData}");
             }
             else
             {
-                Console.WriteLine($"HtmlPdfClient error: {pdfresult.Error!}");
+                Console.WriteLine($"HtmlPdfServer error: {pdfresult.Error}");
             }
-
-            Console.WriteLine("Press any key to next");
+            Console.WriteLine("Press any key");
             Console.ReadKey();
-
-            //create client instance  and send to server
-            Console.WriteLine($"HtmlPdfClient send TemplateRazor to PDF Server via http post");
-
-            var lstprod = new List<Product>();
-            for (int i = 0; i < 40; i++)
-            {
-                lstprod.Add(new Product($"Product{i}", 9.99m));
-            }
-
-            var order1 = new Order("Roberto Rivellino", "Rua S&atilde;o Jorge, 777", "+55 11 912345678", lstprod);
-
-            pdfresult = await HtmlPdfClient.Create("HtmlPdfPlusClient")
-                                 .PageConfig((cfg) => cfg.Margins(10))
-                                 .Logger(HostApp.Services.GetService<ILogger<Program>>())
-                                 .FromRazor(TemplateRazor(), order1)
-                                 .Timeout(5000)
-                                 .Run(clienthttp,applifetime.ApplicationStopping);
-
-            Console.WriteLine($"HtmlPdfClient IsSuccess {pdfresult.IsSuccess} after {pdfresult.ElapsedTime}");
-
-            //performs writing to file after performing conversion
-            if (pdfresult.IsSuccess)
-            {
-                var fullpath = Path.Combine(PathToSamples, "html2pdfRazorTemplate.pdf");
-                await File.WriteAllBytesAsync(fullpath, pdfresult.OutputData!);
-                Console.WriteLine($"File PDF generate at {fullpath}");
-            }
-            else
-            {
-                Console.WriteLine($"HtmlPdfClient error: {pdfresult.Error!}");
-            }
-
-            Console.WriteLine("Press any key to next");
-            Console.ReadKey();
-
-            //create client instance  and send to server
-            Console.WriteLine($"HtmlPdfClient send Url to PDF Server via http post");
-
-            pdfresult = await HtmlPdfClient.Create("HtmlPdfPlusClient")
-                                .PageConfig((cfg) => cfg.Margins(10))
-                                .Logger(HostApp.Services.GetService<ILogger<Program>>())
-                                .FromUrl(new Uri("https://github.com/FRACerqueira/HtmlPdfPlus"))
-                                .Timeout(15000)
-                                .Run(clienthttp, applifetime.ApplicationStopping);
-
-            Console.WriteLine($"HtmlPdfClient IsSuccess {pdfresult.IsSuccess} after {pdfresult.ElapsedTime}");
-
-            //performs writing to file after performing conversion
-            if (pdfresult.IsSuccess)
-            {
-                var fullpath = Path.Combine(PathToSamples, "HtmlPdfPlus.pdf");
-                await File.WriteAllBytesAsync(fullpath, pdfresult.OutputData!);
-                Console.WriteLine($"File PDF generate at {fullpath}");
-            }
-            else
-            {
-                Console.WriteLine($"HtmlPdfClient error: {pdfresult.Error!}");
-            }
-
-
-            Console.WriteLine("Press any key to end");
-            Console.ReadKey();
-
 
         }
-
-        public record Product(string Name, decimal Price);
-
-        public record Order(string CustomerName, string CustomerAddress, string CustomerPhoneNumber, List<Product> Products);
-
         private static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureLogging((hostContext, logbuilder) =>
@@ -147,76 +78,23 @@ namespace ConsoleHtmlToPdfPlus.ClientSendHttp
                         .AddConsole();
                 })
                 .ConfigureServices((hostContext, services) =>
-                { 
-                  services.AddHttpClient("HtmlPdfServer", httpClient =>
-                  {
-                      httpClient.BaseAddress = new Uri("https://localhost:7212/GeneratePdf");
-                  });
+                {
+                    //create an Html2Pdf ServerPlus service with input/output parameter being a file path
+                    services.AddHtmlPdfService<string, string>((cfg) =>
+                    {
+                        cfg.DefaultConfig((cfg) =>
+                            {
+                                cfg.DisplayHeaderFooter(true)
+                                   .Margins(10, 10, 10, 10);
+                            })
+                            .Logger(LogLevel.Debug, "MyPDFServer")
+                            //when run in the same context, not Compress is fast because it is not required to transfer data over the network
+                            .DisableFeatures(DisableOptionsHtmlToPdf.DisableCompress);
+                    });
                 });
 
-        private static string TemplateRazor()
-        {
-            return """
-                <!DOCTYPE html>
-                <html lang="pt-br">
-                <head>
-                    <meta charset="UTF-8">
-                    <title>Customer Details</title>
-                    <style>
-                        table {
-                            border-collapse: collapse;
-                            width: 100%;
-                        }
-                        th, td {
-                            border: 1px solid #ddd;
-                            padding: 8px;
-                        }
-                        th {
-                            background-color: #f4f4f4;
-                            text-align: left;
-                        }
-                        tr { 
-                            page-break-inside: avoid; 
-                        }
-                    </style>
-                </head>
-                <body>
-                    <h1>Customer Details</h1>
-                    <p><strong>Name:</strong> @Model.CustomerName</p>
-                    <p><strong>Address:</strong> @Model.CustomerAddress</p>
-                    <p><strong>Phone Number:</strong> @Model.CustomerPhoneNumber</p>
-
-                    <h2>Products (@Model.Products.Count)</h2>
-                    @if(Model.Products.Any())
-                    {
-                        <table>
-                            <thead>
-                                <tr>
-                                    <th>Product Name</th>
-                                    <th>Price</th>
-                                </tr>
-                            </thead>
-                            <tbody>
-                                @foreach (var product in Model.Products)
-                                {
-                                    <tr>
-                                        <td>@product.Name</td>
-                                        <td>@product.Price.ToString("C")</td>
-                                    </tr>
-                                }
-                            </tbody>
-                        </table>
-                    } 
-                    else
-                    {
-                        <p>No products found.</p>
-                    }
-                </body>
-                </html>
-                """;
-
-        }
-
+        // The [{MyTokenTemplate}] placeholder below is substituted by the BeforePDF hook in
+        // Main before conversion - that substitution is the whole point of this sample.
         private static string HtmlSample()
         {
             return
@@ -226,7 +104,7 @@ namespace ConsoleHtmlToPdfPlus.ClientSendHttp
                 <head>
                     <meta charset="UTF-8">
                     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                    <title>HTML to PDF Test</title>
+                    <title>[{MyTokenTemplate}]</title>
                     <style>
                         body {
                             font-family: Arial, sans-serif;
